Initial commit. Create basic script and functionality
v1.0 - The script will add all subscribed Workshop items to a collection
and then unsubscribe each of them. Steam username and appid of the
game (whose subscribed items will be processed) are configurable inside
the script.
All items tagged as "mod" will be sent to the mods collection, and all other
ones will be sent to the other collection.
Error control is still pending.